refactor: 동적 html 클래스 구조 수정
