+++ src/main/java/constant/Uri.java
@@ -5,6 +5,7 @@ public class Uri {
     public static final String USER_LOGIN_URI = "/user/login.html";
     public static final String USER_LOGIN_FAILED_URI = "/user/login_failed.html";
     public static final String USER_LIST_URI = "/user/list.html";
+    public static final String QNA_FORM_URI = "/qna/form.html";
     public static final String USER_CREATE_REQUEST_URI = "/user/create";
     public static final String USER_LOGIN_REQUEST_URI = "/user/login";
