+++ src/main/java/controller/ControllerImpl.java
@@ -5,19 +5,19 @@ import controller.annotation.PostMapping;
 import db.Database;
 import dto.LoginRequestDto;
 import dto.UserFormRequestDto;
-import dto.UserResponseDto;
 import model.HttpRequest;
 import model.HttpResponse;
 import model.User;
 import model.enums.HttpStatusCode;
 import model.enums.Mime;
-import service.DynamicHtml;
 import service.FileService;
 import service.UserService;
 import session.Authorization;
 import session.HttpSession;
+import view.DynamicHtml;
 
 import java.io.IOException;
+import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 
@@ -37,15 +37,24 @@ public class ControllerImpl implements Controller {
         userService = new UserService();
     }
 
+    @GetMapping
+    public HttpResponse getHttpResponse(HttpRequest httpRequest) throws IOException {
+        String path = httpRequest.getUri();
+        Mime extension = Mime.getValueOf(path);
+
+        byte[] fileContents = fileService.openFile(path, extension);
+        return createHttpResponse(HttpStatusCode.OK, fileContents, extension);
+    }
+
     @GetMapping(path = "/index.html")
     public HttpResponse getIndexHtml(HttpRequest httpRequest) throws IOException {
         String sessionId = httpRequest.getSessionIdInCookie();
         Optional<HttpSession> session = Authorization.getSession(sessionId);
 
-        UserResponseDto userData = null;
-        if (session.isPresent()) userData = session.get().getUserData();
+        Map<String, Object> viewParameters = new HashMap<>();
+        session.ifPresent(httpSession -> viewParameters.put("users", httpSession.getUserData()));
 
-        byte[] content = new DynamicHtml().index(userData);
+        byte[] content = DynamicHtml.select(httpRequest, viewParameters);
         return getHttpResponse(httpRequest.getUri(), content);
     }
 
@@ -56,8 +65,23 @@ public class ControllerImpl implements Controller {
         if (session.isEmpty()) {
             return createRedirectResponse(HttpStatusCode.MOVED_PERMANENTLY, USER_LOGIN_URI);
         }
-        byte[] contents = new DynamicHtml().userList(Map.of("users", Database.findAll()));
+
+        Map<String, Object> viewParameter = new HashMap<>();
+        viewParameter.put("users", Database.findAllUser());
+
+        byte[] contents = DynamicHtml.select(httpRequest, viewParameter);
         return getHttpResponse(httpRequest.getUri(), contents);
+    }
+
+    // FIXME step-7
+    @GetMapping(path = "/qna/form.html")
+    public HttpResponse write(HttpRequest httpRequest) throws IOException {
+        String sessionId = httpRequest.getSessionIdInCookie();
+        Optional<HttpSession> session = Authorization.getSession(sessionId);
+        if (session.isEmpty()) {
+            return createRedirectResponse(HttpStatusCode.MOVED_PERMANENTLY, USER_LOGIN_URI);
+        }
+        return getHttpResponse(httpRequest);
     }
 
     @PostMapping(path = "/user/login")
@@ -77,14 +101,6 @@ public class ControllerImpl implements Controller {
     public HttpResponse addUserByForm(Map<String, String> body) {
         userService.createByForm(new UserFormRequestDto(body));
         return createRedirectResponse(HttpStatusCode.MOVED_PERMANENTLY, INDEX_HTML_URI);
-    }
-
-    @GetMapping
-    public HttpResponse getHttpResponse(HttpRequest httpRequest) throws IOException {
-        String path = httpRequest.getUri();
-        Mime extension = Mime.getValueOf(path);
-        byte[] fileContents = fileService.openFile(path, extension);
-        return createHttpResponse(HttpStatusCode.OK, fileContents, extension);
     }
 
     private HttpResponse getHttpResponse(String path, byte[] body) throws IOException {
