+++ src/main/java/db/Database.java
@@ -12,10 +12,6 @@ import java.util.stream.Collectors;
 public class Database {
     private static Map<String, User> users = Maps.newHashMap();
 
-    // 테스트용 데이터
-    static {
-        users.put("qq", new User("qq", "qq", "kim", "qq@qq.qq"));
-    }
 
     public static void addUser(User user) {
         users.put(user.getUserId(), user);
@@ -25,7 +21,7 @@ public class Database {
         return Optional.ofNullable(users.get(userId));
     }
 
-    public static List<UserResponseDto> findAll() {
+    public static List<UserResponseDto> findAllUser() {
         return users.values()
                 .stream()
                 .map(UserResponseDto::of)
