+++ src/main/java/view/Servlet.java
@@ -1,0 +1,7 @@
+package view;
+
+import java.util.Map;
+
+public interface Servlet {
+    byte[] doGet(Map<String, Object> data);
+}
